// src/services/workflow/dtos/workflow-dto.ts
export interface WorkflowDTO {
    workflowId: string;
    name: string;
    description?: string;
    status: string;
    type: string;
    config?: any;
    metadata?: Record<string, unknown>;
    tags: string[];
    createdBy?: string;
    createdAt: string;
    updatedAt: string;
    isDeleted: boolean;
}